third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/interactions-of-styling-and-scripting/style-element-media-match-block-script.html

<!DOCTYPE html>
<title>Style element is script-blocking when media matches</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const start = performance.now();
</script>
<style>
@import url('stylesheet.py?delay=1');
</style>
<script>
// The loading of the imported sheet may start before the <style> element is set
// up due to preload scanner, so even when it's script-blocking, it may block
// script for a bit less than 1000ms. So we set an error margin to make the test
// more robust.
const epsilon = 50;
test(() => {
  const delay = performance.now() - start;
  assert_greater_than(delay, 1000 - epsilon);
});
</script>
